refactor: 팀 플레이어 관련 view 분리
