+++ src/sportslive/team/presentation/__init__.py
@@ -1,5 +1,6 @@
 from .team_view import TeamView
 from .team_get_view import TeamGetView
-from .team_player_view import TeamPlayerView
+from .team_player_create_view import TeamPlayerCreateView
 from .team_player_get_view import TeamPlayerGetView
 from .team_register_view import TeamRegisterView
+from .team_player_update_delete_view import TeamPlayerUpdateDeleteView
