+++ src/sportslive/team/presentation/team_player_update_delete_view.py
@@ -6,11 +6,10 @@ from accounts.domain import IsAdminUser
 from team.containers import TeamContainer
 from drf_yasg.utils import swagger_auto_schema
 from team.serializers import (
-    TeamPlayerRegisterRequestSerializer,
     TeamPlayerChangeRequestSerializer
 )
 
-class TeamPlayerView(APIView):
+class TeamPlayerUpdateDeleteView(APIView):
     authentication_classes = [JWTAuthentication]
     permission_classes = [IsAdminUser]
 
@@ -18,14 +17,6 @@ class TeamPlayerView(APIView):
         super().__init__(*args, **kwargs)
         self._team_player_service = TeamContainer.team_player_service()
 
-
-    @swagger_auto_schema(request_body=TeamPlayerRegisterRequestSerializer(many=True), responses={"201": ''})
-    def post(self, request, team_id: int):
-        """
-        리그팀 플레이어 생성 API
-        """
-        self._team_player_service.register_team_players(request_data=request.data, team_id=team_id)
-        return Response(status=status.HTTP_201_CREATED)
     @swagger_auto_schema(request_body=TeamPlayerChangeRequestSerializer, responses={"200": ''})
     def put(self, request, team_player_id: int):
         """
